Update The Witness.yaml
Some choices I made:

`Vague hints: Experimental` is completely fine & stable, but I think it adds a bit more chaos than necessary for the big async.

`victory_condition: panel_hunt` is given a slightly higher weight than other victory conditions. This is because the other victory conditions are very similar, especially the two box goals. Also, I think panel hunt will be very fun for newcomers, it has been incredibly well received by the playerbase. I was actually also debating lowering the weights of `mountain_box_short` and `mountain_box_long` because they are so similar, but ultimately decided against it for simplicity.

`panel_hunt_postgame` is limited to the two simplest options, it is already confusing enough as is.

`puzzle_randomization` now has a 20% chance for the new `umbra_variety` mode, and also a 10% chance for expert randomizer, as there are a few "expert heads" who would be sad if there are no expert seeds to play in the new filler-slot-only world.