Add Grub Hunt to HK (#218)
* Add Grub Hunt to HK

Adds Grub Hunt to HK. If the goal is grub hunt, the required number of grubs is a random range from 25 to 35, and a trigger is set up to make grubs local.

* add trigger for Any goal.yaml

Due to unnoticed behaviour regarding The "Any" option for goal, the choice was either to create a second trigger for the Any goal to make grubs local, or to remove Any as an option. I feel as if I saw enough positive opinions on Any goal being like this that I went with the former.